Changed ControllerInput, Created TextMain, Organized Display
Changed ControllerInput class's name to UserInput, Changed charMovementInput() to movementInput(), in movementInput(): returns a Position object instead of a int[], includes a main method for testing inputs.

Created TextMain class, includes a skeleton of validInput(), and a basic main method using the code within Demo1, does not compile at the moment, needs a fixed Display.

Fixed a bit of coding syntax within Display, put in the codes outside of printGrid into the method, and removed the instantiation of Map, a map object is now instantiated within TextMain and so DIsplay should call for  the map from there, did not fix the functionality of this class.

Wrote some basic descriptions for TextMain and UserInput.